backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Modular sales and inventory management system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.1",
    "pydantic-settings>=2.1.0",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "httpx>=0.26.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==3.2.2",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
